+++ src/api/BaseHttpHandler.java
@@ -1,4 +1,4 @@
-package HttpServer;
+package api;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
