+++ src/api/DurationTypeAdapter.java
@@ -1,4 +1,4 @@
-package HttpServer;
+package api;
 
 import com.google.gson.TypeAdapter;
 import com.google.gson.stream.JsonReader;
