+++ src/api/Endpoint.java
@@ -1,4 +1,4 @@
-package HttpServer;
+package api;
 
 public enum Endpoint {
     GET_TASK, GET_TASKS, POST_TASK, DELETE_TASK, GET_SUBTASK, GET_SUBTASKS, POST_SUBTASK, DELETE_SUBTASK, GET_EPIC, GET_EPICS, POST_EPIC, DELETE_EPIC, GET_EPICS_SUBTASKS, HISTORY, PRIORITIZED, UNKNOWN
